fix(f16): accumulate the scalar conv/matmul f16 legs in float32
The scalar (non-MVE and ARM_MATH_AUTOVECTORIZE) f16 legs of arm_convolve_f16's
OHWI/NT_N_PACKED fallback, arm_nn_mat_mult_nt_t_f16 and
arm_nn_mat_mult_nt_n_packed_f16 accumulated in _Float16, drifting 3-4 f16 ulp
over a 27-tap chain against the tester's exact-value nonfinite conv cases
(helia-core-tester v0.2.51, m55-f16-fallback leg: lanes 591/821/873 of
convolve_float_nonfinite_{nan,inf,neginf}_f16). Widen bias and every product to
float32 and round to f16 once at the store, as arm_reduce_sum_f16 and the
conv1d k3/k5 scalar legs already do; the f16 clamp after that single rounding
keeps the NaN/Inf contract byte-for-byte. The MVE legs are untouched (m55
-Ofast objects byte-identical) and their float16 lane accumulation is now
stated per leg in the header notes.

Refs #449, refs #457

+++ Source/NNSupportFunctions/arm_nn_mat_mult_nt_n_packed_f16.c
@@ -168,12 +168,13 @@ arm_cmsis_nn_status arm_nn_mat_mult_nt_n_packed_f16(const float16_t *__RESTRICT 
                 vst1q_p(dst_row + c, vacc, p);
             }
     #else
+            /* Scalar leg: float32 accumulation, one f16 rounding before the clamp (#449, #457). */
             for (int32_t lane = 0; lane < valid_cols; ++lane)
             {
-                _Float16 acc = bias ? (_Float16)bias[c + lane] : (_Float16)0.0f;
+                float32_t acc = bias ? (float32_t)bias[c + lane] : 0.0f;
                 for (int32_t k = 0; k < rhs_cols; ++k)
                 {
-                    acc += (_Float16)lhs_row[k] * (_Float16)rhs_block[(size_t)k * block_cols + lane];
+                    acc += (float32_t)lhs_row[k] * (float32_t)rhs_block[(size_t)k * block_cols + lane];
                 }
                 dst_row[c + lane] = arm_nn_clamp_scalar_f16((float16_t)acc, activation_min, activation_max);
             }
